src/requirements.txt
annotated-types==0.7.0
argcomplete==3.6.3
black==26.1.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
datamodel-code-generator==0.53.0
genson==1.3.0
idna==3.11
inflect==7.5.0
isort==7.0.0
Jinja2==3.1.6
MarkupSafe==3.0.3
more-itertools==10.8.0
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.6.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
pytokens==0.4.1
PyYAML==6.0.3
requests==2.32.5
typeguard==4.4.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
